# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TASK CALENDAR",
        "description": "Put Markdown Here [a Link](https://www.google.com)",
        "contact": {
            "email": "my@email"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/task": {
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Criando nova tarefa",
                "operationId": "752c97798f6581341105be3afd0d061a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assigneduser_id",
                                    "title",
                                    "description",
                                    "due",
                                    "completed"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "assigneduser_id": 1,
                                    "title": "New task",
                                    "description": "loren impsum",
                                    "due": "2024-04-11",
                                    "completed": false
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": true,
                                            "message": "Task Created Successfully",
                                            "task": {
                                                "assigneduser_id": 1,
                                                "title": "New task",
                                                "description": "loren impsum",
                                                "due": "2024-04-11",
                                                "completed": 0,
                                                "taskcreator_id": 1,
                                                "updated_at": "2024-04-16T00:40:56.000000Z",
                                                "created_at": "2024-04-16T00:40:56.000000Z",
                                                "id": 8
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": false,
                                            "message": "validation error",
                                            "errors": {
                                                "fields": [
                                                    "The assigneduser_id not found."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": false,
                                            "message": "Throwable errors"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Pegando todas as tarefas",
                "description": "Para o retorno dos dados, deve-se autenticar e pegar o token gerado. Clicar no cadeado e colocar Bearer mais o token gerado no login. Ex: Bearer 12|KvkT1313lSelWBH5xXxokLYOhN38B3XsS3riNuLKcca49988",
                "operationId": "8549bd34c3368db59870d2aeb3e8fe2a",
                "responses": {
                    "200": {
                        "description": "Get all successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "id": 1,
                                            "taskcreator_id": 1,
                                            "assigneduser_id": 1,
                                            "title": "New task",
                                            "description": "loren impsum",
                                            "due": "2024-04-11 00:00:00",
                                            "completed": 0,
                                            "created_at": "2024-04-15T20:24:28.000000Z",
                                            "updated_at": "2024-04-15T20:24:28.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": false,
                                            "message": "Throwable errors"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/task/{task}/finished": {
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Close a task",
                "operationId": "9985ae003eeef4f716afcc59963dd90d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da task",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}